refactor : 테스트 코드에서 ThresholdBaseMoveDecider, GreaterThanOrEqualThreshold 구현체 사용하도록 변경
